Always use custom fetch
