+++ packages/server/src/sdk/workspace/ai/agents/utils.ts
@@ -16,7 +16,6 @@ import {
 } from "../../../../ai/tools"
 import sdk from "../../.."
 import { createExaTool, createParallelTool } from "../../../../ai/tools/search"
-import tracer from "dd-trace"
 
 export function toToolMetadata(tool: AiToolDefinition): ToolMetadata {
   return {
@@ -167,50 +166,6 @@ function addHelperTools(
   }
 
   return enabledTools
-}
-
-export function createLiteLLMFetch(sessionId: string): typeof fetch {
-  const liteFetch = (async (
-    input: Parameters<typeof fetch>[0],
-    init?: Parameters<typeof fetch>[1]
-  ) => {
-    const span = tracer.scope().active()
-    let modifiedInit = init
-
-    if (typeof init?.body === "string") {
-      try {
-        const body = JSON.parse(init.body)
-        body.litellm_session_id = sessionId
-        if (span) {
-          body.metadata = {
-            ...body.metadata,
-            dd_trace_id: span.context().toTraceId(),
-            dd_span_id: span.context().toSpanId(),
-            session_id: sessionId,
-          }
-        }
-        modifiedInit = { ...init, body: JSON.stringify(body) }
-      } catch {
-        // Not JSON, pass through
-      }
-    }
-
-    const response = await fetch(input, modifiedInit)
-
-    const litellmCallId = response.headers.get("x-litellm-call-id")
-    if (litellmCallId && span) {
-      span.setTag("litellm.call_id", litellmCallId)
-    }
-
-    return response
-  }) as typeof fetch
-
-  // Preserve the preconnect helper required by the OpenAI client typings.
-  if (typeof (fetch as any).preconnect === "function") {
-    ;(liteFetch as any).preconnect = (fetch as any).preconnect.bind(fetch)
-  }
-
-  return liteFetch
 }
 
 export interface IncompleteToolCall {
